requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
azure-functions
azure-cosmos  # Still needed for local Cosmos DB operations
openai
pytz
pytest
pytest-asyncio
python-dotenv
